+++ data_engineering/derive_pipeline_runner.py
@@ -1,8 +1,6 @@
-import pandas as pd
-
 from utils import pipeline_io
 
-from data_engineering.derive.derive import get_raw_values
+from data_engineering.utils.derive import get_raw_values
 
 
 def run():
